Do less in worker
